change!: require `Ancestors` traversal `find()` to return `Result` (#364)
Previously it would return an `Option` and squelch legitimate errors
which already let to quite a bug hunt.

Now `find()` should always return the original error when finding an
existing object, and keep it around in a Box.
That way we don't add yet another type parameter.

Performance implications are measurable, but marginable. The error type
might be too big which is something we can indeed look at.