empty crate for git-note
It came up when adding categorization of references and it's something
that should be useful one day for additional tools, maybe even as
foundation for git-powered issues or something.

In any case, implementing it will help get acquainted with storing
metadata in git which certainly helps for doing some sort of database
in git.