Support for shared attribute file names (#301)
This isn't much of a performance win and is mostly there to prove
how we can handle the intricacies of a separate path backing
in the light of necessary mutations.

Temporarily splitting pieces out of the index seems to be the way
to go even though it's not super easy to implement, but it makes
possible what we have to do.