Allow check-ignore style queries with API that doesn't remove trailing slashes (#301)
It's really only for check-ignore and for our tests as these don't
happen when working with the index or the file system.

However, we are dependent on the order of operations and assume properly
ordered input which can't be assumed for things like check-ignore
and can probably lead to strange bugs if some users don't know about
that.

There should be a fix for that.