`path::discover()` now returns the shortest path. (#301)
If and only if it canonicalized the source path. That way, users will
still get a familiar path. This is due to `parent()` not operating
in the file system, which otherwise would be equivalent to `..`,
but that's not how we work.

Maybe we should overhaul the way this works to use `../` instead
and just 'absoluteize' the path later (std::path::absolute()) is
on the way for that.