reorganize types to properly represent worktrees in their various 'states' (#301)
We can iterate them, as proxies, and put them into a support type
to make it easy to assert we actually have a worktree.

Besides that, worktrees are really just another facet of a repository,
but we chose to implement their functionality on a `Worktree` to
avoid having to check for the lack of worktree each time we do
something involving a worktree.