hopefully fix msrv check and bring tools back into workspace
Note that we must not duplicate the MSRV version to check as it's
currently controlled by the `msrv.yml` template. For local testing,
keep command-lines local instead of adjusting the makefile.

For now, we have to skip the check as it just doesn't work with
this rustc version at least.