Split objects stores
Instead of storing all objects in one big object store, this will
keep mirrored (objects that are found on the upstream) and computed
(objects that are produced by filtering) in separate git object
databases.
The git alternates mechanism is then used to make objects available
to the other repo when needed.

The reason for this change is a very significant performance
degradation during `git fetch` on very large repos. The huge amount
of loose objects slowed down `git index-pack` (which is run after
downloading the packfile). With this change all the loose objects
produced by josh filtering are not visible to `index pack` and
thus cannot affect it's performance anymore.

Change-Id: split-odb