change!: Simplify `Boolean` to be a wrapper around `bool`. (#331)
Previously it tried hard not to degenerate information, making it a
complicated type.

However, in practice nobody cares about the exact makeup of the boolean,
and there is no need to serialize a boolean faithfully either.

Instead, those who want to set a value just set any value as a string,
no need for type safety there, and we take care of escaping values
properly on write.