change!: Associate `file::Metadata` with each `File`. (#331)
This is the first step towards knowing more about the source of each
value to filter them based on some properties.

This breaks various methods handling the instantiation of configuration
files as `file::Metadata` typically has to be provided by the caller
now or be associated with each path to read configuration from.