add support for keeping multiple candidates in case of ambiguous objects. (#427)
Now, test for test, we can work through all the cases that now fail to
decide what needs to be done or implemented to make them work properly.

Seems like quite some work, but rewarding at such as it will be possible
to neatly disambiguate everything automatically and surpass git in
capabilities naturally when it comes to this.