fix: always update modification date when packed refs are forcefully reloaded. (#427)
The latter happens right after the file was written, which means we
definitely have to update our modification date or else the buffer
will be read again next time somebody asks for an up-to-date version.

The ordering of operations is such that at worst, the buffer is loaded
again instead of keeping an outdated version of it, which would happen
if the metadata is read afterwards.