cli: hide `jj close/commit` and `jj open/uncommit`
Now that open commits are disabled by default, seeing `close` and
`open` in `jj help` is confusing.

I'm a little ambivalent about hiding `jj commit` because it's useful
as a way of adding a description and starting a new change with a
single command. We can always add a new command for that if we want
to.