fix: remove `Drop` for `SpawnProcessOnDemand`. (#450)
It is well-intended but is likely to hang the calling process
if for any reason not all process output was consumed.

Even though not waiting for the process leaves it running, it will
stop naturally once its output pipe breaks once once
our handles for it are inevitable dropped at the same time.