no need to fetch the whole pack in dry-run mode, even in V1. (#450)
This demands explanation. Previously the issue was a blocking transport
implementation on drop, but that has been fixed by simply letting the
child process run into a broken pipe to stop it.

Now existing early won't hang because of the transport being dropped
with data still available on the stdout pipe.

The issue to fix here is that keep-alive packets are not currently
understood in V1, and git can send us empty packs.