Add test to validate empty fetches are OK
This can happen if the negotiation ends up requesting objects that
it also claims to have, with no object actually being required to be
sent after all.

This could be fixed on negotation level, but since NAIVE negotiations
are prone to change and maybe are replaced in favor of supporting the
official algorithms, there is no use in trying to assure it won't
request empty packs.