fix: reduce memory usage for pack resolution by 18%…
…compared to where it started out before.

This is the result of using u32 intead of usize where possible,
leveraging the object limit of u32 in packs.

This change makes the biggest difference by not keeping the decompressed
memory of leaf nodes alive for longer than needed, at the cost of
some code duplication which could be extracted into a function if one
was inclined to deal with the boilerplate and lots of generics.